+++ example/main.dart
@@ -9,13 +9,21 @@ Future<void> main(List<String> args) async {
   // get specific browser
   var browser = Chrome(sqlite3Path: './sqlite3.dll');
 
-  // get history from this browser
-  var histories = await browser.fetchHistory();
+  // get profiles
+  var profiles = browser.fetchProfiles();
+  print(profiles);
 
-  var bookmarks = await browser.fetchBookmarks();
+  // get history from browser
+  var histories = await browser.fetchHistory(profiles: ['Profile 1']);
+  print(histories);
 
-  // get bookmarks from this browser
+  // get bookmarks from browser
+  var bookmarks = await browser.fetchBookmarks(profiles: ['Default']);
+  print(bookmarks);
+
+  // get passwords from browser
   var passwords = await browser.fetchPasswords();
+  print(passwords);
 }
 
 void formatPrint(Map<String, dynamic> map) {